PaperPR Cache resource keys
Setting biomes using Bukkit's `World#setBiome`, delegating to CraftWorld#setBiome will create a new ResourceKey every time a biome is set, using String#intern on the biome ID obtained from the bukkit biome, convert it to an internal BiomeBase. This is an extremely resource-costly method and setting relatively small regions of biome (in this example a 200x200x384 region) can have a massive performance impact on the server.